Merged PR 1269: Fix for error messages in the log:
Fix for error messages in the log:

Unable to parse any XML input. Error on line 1: The prefix "p" for attribute "p:before" associated with an element type "r" is not bound

Not sure who added this XML,. but it was throwing errors in the log, publishing failed. Had a look at HabitatHome home item, and changed the namespace like it was there.... problem solved, Sitecore also shows the correct UI in terms of the renderings again, which it didn't before

Related work items: #2764