Merged PR 2785: Issue with multiple form submissions - reset to anonymous contact in demo sidebar
Updating method to identify contact - if contact is already known and identified then do not call identifyAs.

The result of this bug fix is that multiple form fills do not lead to an anonymous contact shown in the demo sidebar. Please compare against v10 master (stable) to see the issue with multiple form fills.

As part of this bug fix, the First name, Last name, Phone number is not displayed in the demo sidebar in real time after a form fill. A different task will be created to investigate that functionality since it has conflicting functionality.

Related work items: #5393