Merged PR 2879: Update Content Hub Entity ID for Articles
Added a new CMP mapping in Sitecore. The Content Hub Entity ID for Articles as it was changed in the Lighthouse CH template. This change will allow old and new CH instances to work.

The new Entity ID was introduced in the Lighthouse template for CH 3.4.2 and above.

See testing instructions in attached bug.

Related work items: #5534