Merged PR 2898: #5436 Fixing facet reading
It seems to be the only solution that keeps the data in order.

I tried to use Tracker.Current.Session.Contact = contactManager.LoadContact(Tracker.Current.Contact.ContactId); but it definitely breaks adding user to Automation Plans if executed on the same request when it is being added. Looks like a Sitecore bug, but needs more time for proper reproducing.

Related work items: #5436