Merged PR 3398: Add comment to XML override file
We are overriding the OOTB personalization dialog so the "Track Personalization Effect" checkbox is not set by default (result is personalization is not shown 10% of the time if this checkbox is set).

The Sitecore AI Standard module introduced another personalization dialog override file and we need to override it as well. It is override inception.

Related work items: #5964