Merged PR 3228: Enable Auto content tagging
adding  OPENCALAIS_ACCESSTOKEN to docker compose arg list so it gets passed thru to build

pull3228-contenttagging instance is ready but tagging is not working by the way

I thought the new way of passing in the variable would work on here as it was throwing 401 unauthorized on dev machine. but It's not working.
https://www.loom.com/share/8835336d6228472a85f15e9ed64635b9

Related work items: #5818