Merged PR 3574: Update SOLR_CORE_PREFIX_NAME Setting for CM/CD Roles
The SOLR_CORE_PREFIX_NAME setting was hardcoded for CM/CD in docker-compose.override file - this change makes the setting read dynamically as expected. Setting is already in docker-compose file for CM role. We need to add for CD in override file.

Note that the SOLR_CORE_PREFIX_NAME setting does not yet work with SXA out-of-the-box but the configs should be updated in 10.2:
Product task: https://dev.azure.com/Sitecore-PD/Products/_workitems/edit/464193

Related work items: #6105