Merged PR 3540: Lighthouse Healthcare Vertical
Please review the Healthcare site available at the following virtual URL: **/healthcare**

The only remaining notes I gave to Dan/Wendy is that the following 2 pages are empty. They may decide to add demo content but it is not part of the demo flow for this site and not necessary at this time.

Empty pages:
 - /healthcare/Medical-Professionals
 - /healthcare/Visitors-and-Families

Other completed tasks:
 - Added pages to the warmup script for the new site
 - Dynamically setting the Base URL for the EXM root in the new site
 - Run Broken links/reference report for new items

While I'm gone on PTO, please feel free to make updates as required or complete this PR to get it into the develop branch. We can do a final confirmation with Dan/Wendy to make sure there are no final changes and can get this into the master branch so they can share post about the new Healthcare vertical site.

Related work items: #6086